IN PROGRESS - issue CORE-4152: Implement App Integration WYSIWYG Pallete Drag n' Drop https://pods.iplantcollaborative.org/jira/browse/CORE-4152
Added DragSource capability to the images within AppIntegrationPalette.
Now, on drag start, the drag source event data is loaded with the
ArgumentType associated with the image being dragged. This allows any
drop targets who are in the know have knowledge of what argument type
control is intended to be added to the App wizard widget.